+++ module/controllers/healing-roll.js
@@ -1,24 +1,26 @@
 export class CofHealingRoll {
-    constructor(label, formula, isCritical, title, showButtons=true, description){
+    constructor(label, formula, isCritical, title, showButtons=true, description, rollMode){
         this._label = label;
         this._formula = formula;
         this._isCritical = isCritical;
         this._title = title;
         this._showButtons = showButtons;
         this._description = Array.isArray(description) ? description.join("<br>") : description;
+        this._rollMode = rollMode;
     }
 
     async roll(actor){
         const r = new Roll(this._formula,actor.system);
         await r.roll();
         if (this._isCritical) r._total = r._total * 2;
+        const messageOptions = { rollMode: this._rollMode };
         this._buildHealingRollMessage().then(msgFlavor => {
             r.toMessage({
                 user: game.user.id,
                 flavor: msgFlavor,
                 speaker: ChatMessage.getSpeaker({actor: actor}),
                 flags : {msgType : "heal"}
-            });
+            }, messageOptions);
         });
         return r;
     }