+++ module/controllers/roll.js
@@ -335,25 +335,31 @@ export class CofRoll {
    * @param {*} description
    * @returns
    */
-  static async skillRollDialog(actor, label, mod, bonus, malus, critrange, superior = false, onEnter = "submit", description, weakened = false) {
+  static async skillRollDialog(actor, label, mod, bonus, malus, critrange, superior = false, onEnter = "submit", description, weakened = false, dice, rollMode, difficulty) {
     const rollOptionTpl = "systems/cof/templates/dialogs/skillroll-dialog.hbs";
     let diff = null;
     const displayDifficulty = game.settings.get("cof", "displayDifficulty");
-    if (displayDifficulty !== "none" && game.user.targets.size > 0) {
-      diff = [...game.user.targets][0].actor.system.attributes.def.value;
-    }
+    if (!foundry.utils.isEmpty(difficulty)) diff = difficulty;
+    else {
+      if (displayDifficulty !== "none" && game.user.targets.size > 0) {
+        diff = [...game.user.targets][0].actor.system.attributes.def.value;
+      }
+    }    
     const isDifficultyDisplayed = displayDifficulty === "all" || (displayDifficulty === "gm" && game.user.isGM);
     const rollOptionContent = await renderTemplate(rollOptionTpl, {
-      mod: mod,
-      bonus: bonus,
-      malus: malus,
-      critrange: critrange,
+      mod,
+      bonus,
+      malus,
+      critrange,
       difficulty: diff,
       displayDifficulty: isDifficultyDisplayed,
-      superior: superior,
+      superior,
       hasDescription: description && description.length > 0,
       skillDescr: description,
-      weakened: weakened,
+      weakened,
+      dice,
+      rollMode,
+      rollModes: CONFIG.Dice.rollModes
     });
     let d = new Dialog(
       {
@@ -375,7 +381,8 @@ export class CofRoll {
               const mod = html.find("input#mod").val();
               const bonus = html.find("input#bonus").val();
               const malus = html.find("input#malus").val();
-              let r = new CofSkillRoll(label, dice, mod, bonus, malus, difficulty, critrange, description);
+              const rollMode = html.find("#rollMode").val();
+              let r = new CofSkillRoll(label, dice, mod, bonus, malus, difficulty, critrange, description, rollMode);
               r.roll(actor);
             },
           },
@@ -400,7 +407,7 @@ export class CofRoll {
    * @param {*} onEnter
    * @returns
    */
-  static async rollWeaponDialog(actor, label, mod, bonus, malus, critrange, dmgFormula, dmgBonus, onEnter = "submit", skillDescr, dmgDescr, difficulty = null, weakened = false) {
+  static async rollWeaponDialog(actor, label, mod, bonus, malus, critrange, dmgFormula, dmgBonus, onEnter = "submit", skillDescr, dmgDescr, difficulty = null, weakened = false, rollMode = "publicroll") {
     const rollOptionTpl = "systems/cof/templates/dialogs/roll-weapon-dialog.hbs";
     let diff = null;
     let isDifficultyDisplayed = true;
@@ -430,6 +437,8 @@ export class CofRoll {
       hasDmgDescr: dmgDescr && dmgDescr.length > 0,
       dmgDescr: dmgDescr,
       weakened: weakened,
+      rollMode: rollMode,
+      rollModes: CONFIG.Dice.rollModes
     });
 
     let d = new Dialog(
@@ -453,10 +462,11 @@ export class CofRoll {
               const bonus = html.find("input#bonus").val();
               let malus = html.find("input#malus").val();
               if (!malus) malus = 0;
+              const rollMode = html.find("#rollMode").val();
 
               // Jet d'attaque uniquement
               if (!game.settings.get("cof", "useComboRolls")) {
-                let r = new CofSkillRoll(label, dice, mod, bonus, malus, diff, critrange, skillDescr);
+                let r = new CofSkillRoll(label, dice, mod, bonus, malus, diff, critrange, skillDescr, rollMode);
                 r.weaponRoll(actor, "", dmgDescr);
               } else {
                 // Jet combiné attaque et dommages
@@ -477,8 +487,8 @@ export class CofRoll {
                     dmgFormula = dmgFormula.concat(" ", dmgBonus);
                   }
                 }
-                let r = new CofSkillRoll(label, dice, mod, bonus, malus, diff, critrange, skillDescr);
-                r.weaponRoll(actor, dmgFormula, dmgDescr);
+                let r = new CofSkillRoll(label, dice, mod, bonus, malus, diff, critrange, skillDescr, rollMode);
+                r.weaponRoll(actor, dmgFormula, dmgDescr, rollMode);
               }
             },
           },
@@ -491,7 +501,7 @@ export class CofRoll {
     return d.render(true);
   }
 
-  static async rollDamageDialog(actor, label, formula, dmgBonus, critical = false, onEnter = "submit", dmgDescr) {
+  static async rollDamageDialog(actor, label, formula, dmgBonus, critical = false, onEnter = "submit", dmgDescr, rollMode) {
     const rollOptionTpl = "systems/cof/templates/dialogs/roll-dmg-dialog.hbs";
     const rollOptionContent = await renderTemplate(rollOptionTpl, {
       dmgFormula: formula,
@@ -500,6 +510,8 @@ export class CofRoll {
       isCritical: critical,
       hasDescription: dmgDescr && dmgDescr.length > 0,
       dmgDescr: dmgDescr,
+      rollMode: rollMode,
+      rollModes: CONFIG.Dice.rollModes
     });
 
     let d = new Dialog(
@@ -535,7 +547,8 @@ export class CofRoll {
                 }
               }
 
-              let r = new CofDamageRoll(label, dmgFormula, isCritical, dmgDescr);
+              const rollMode = html.find("#rollMode").val();
+              let r = new CofDamageRoll(label, dmgFormula, isCritical, dmgDescr, rollMode);
               r.roll(actor);
             },
           },